requirements-perf.txt
# Performance testing requirements for create-project
# Install with: uv pip install -r requirements-perf.txt

# Core performance testing
pytest-benchmark>=5.1.0
pytest-profiling>=1.7.0

# Memory profiling
memory-profiler>=0.61.0
psutil>=7.0.0

# CPU profiling
py-spy>=0.4.0
py-cpuinfo>=9.0.0

# Visualization and reporting
matplotlib>=3.5.0
pandas>=1.3.0

# Load testing
pytest-stress>=1.0.1
pytest-repeat>=0.9.3

# Additional tools
objgraph>=3.6.0  # Memory leak detection
line-profiler>=4.0.0  # Line-by-line profiling
snakeviz>=2.1.0  # Profiling visualization